added gitignore, dataset_pazienti and updated main file anaemia_classificator with modules extractor and segmentation
